Added async and await functions

+++ server/server.js
@@ -63,24 +63,23 @@ app.get('/todos/:id', authenticate, (req, res) => {
 });
 
 
-app.delete('/todos/:id', authenticate, (req, res) => {
-    var id = req.params.id; 
-    if(!ObjectID.isValid(id)){
+app.delete('/todos/:id', authenticate, async (req, res) => {
+    try {
+        var id = req.params.id;
+        if(!ObjectID.isValid(id)){
+            return res.status(404).send();
+        } 
+    const todo = await Todo.findOneAndRemove({_id: id, _creator: req.user._id});
+    if(!todo) {
         return res.status(404).send();
     }
 
-    Todo.findOneAndRemove({
-        _id: id,
-        _creator: req.user._id
-    }).then((todo) => {
-        if(!todo) {
-            return res.status(404).send();
-        }
+    res.send({todo});
 
-        res.send({todo});
-    }, (e) => {
+
+    } catch (e) {
         res.status(400).send();
-    });
+    }
 });
 
 
@@ -118,17 +117,16 @@ app.get('/users', (req, res) => {
 
 
 
-app.post('/users', (req, res) => { //SIGN_UP WORKS
-    var body = _.pick(req.body, ['email','password']);
-    var user = new User(body);
-   
-    user.save().then(() => {
-        return user.generateAuthToken();
-    }).then((token) => {
+app.post('/users', async (req, res) => { //SIGN_UP WORKS
+    try {
+        var body = _.pick(req.body, ['email','password']);
+        var user = new User(body);
+        await user.save()
+        const token = await user.generateAuthToken();
         res.header('x-auth', token).send(user);
-    }).catch((e) => {
-        res.status(400).send(e);
-    })
+    } catch (e) {
+         res.status(400).send(e);
+    }
 });
 
 app.get('/users/me', authenticate, (req, res) => {
@@ -149,27 +147,26 @@ app.get('/users/me', authenticate, (req, res) => {
 
 });
 
-app.delete('/users/me/token', authenticate, (req, res) => {
-    req.user.removeToken(req.token).then(() => {
+app.delete('/users/me/token', authenticate, async (req, res) => {
+    try {
+        await req.user.removeToken(req.token);
         res.status(200).send();
-    }, () => {
+    } catch (e) {
         res.status(400).send();
-    });
+    }
 });
 
 
 
-app.post('/users/login', (req, res) => {
-    var body = _.pick(req.body, ['email','password']);
-   
-    User.findByCredentials(body.email, body.password).then((user) => {
-        // res.send(user);
-        return user.generateAuthToken().then((token) => {
-            res.header('x-auth', token).send(user);
-        });  
-    }).catch((e) => {
-        res.status(400).send();
-    });
+app.post('/users/login', async (req, res) => {
+    try {
+        const body = _.pick(req.body, ['email','password']);
+        const user = await User.findByCredentials(body.email, body.password);
+        const token = await user.generateAuthToken();
+        res.header('x-auth', token).send(user);
+    } catch (e) {
+         res.status(400).send();
+    }
 });
